+++ hashcathelper/subcommands/analytics.py
@@ -1,5 +1,8 @@
+import logging
+
 from hashcathelper.args import subcommand, argument
 
+log = logging.getLogger(__name__)
 args = []
 
 args.append(argument(
@@ -36,7 +39,7 @@ case insensitive.
 
 args.append(argument(
     '-f', '--format',
-    choices=['text', 'json', 'html'],
+    choices=['text', 'json', 'html', 'xlsx'],
     default='text',
     help="output format (default: %(default)s)",
 ))
@@ -71,10 +74,57 @@ def analytics(args):
     if not report:
         exit(1)
 
-    out = report.export(args.format)
-
-    if args.outfile:
-        with open(args.outfile, 'w') as f:
-            f.write(out)
+    if args.format == 'xlsx':
+        # xlsx is a bit special because it only contains the details
+        if not args.outfile:
+            log.critical("XLSX format requires OUTFILE to be specified.")
+            exit(1)
+        if not args.degree_of_detail > 2:
+            log.critical(
+                "XLSX format requires degree of detail greater than 2."
+            )
+            exit(1)
+        save_to_xlsx(report, args.outfile)
     else:
-        print(out, end='')
+        out = report.export(args.format)
+
+        if args.outfile:
+            with open(args.outfile, 'w') as f:
+                f.write(out)
+        else:
+            print(out, end='')
+
+
+def save_to_xlsx(report, path):
+    """Saves 'details' from the report to a spreadsheet"""
+
+    from collections import OrderedDict
+    import openpyxl as pyxl
+    from hashcathelper.consts import labels
+
+    workbook = pyxl.Workbook()
+    data = report._elements['details']._elements
+
+    offset = 3
+
+    for k, v in data.items():
+        ws = workbook.create_sheet(k)
+        cell = ws.cell(1, 1, labels.get(k, k))
+        cell.font = pyxl.styles.Font(bold=True, size=14)
+
+        if isinstance(v, (list, tuple)):
+            for i, each in enumerate(v):
+                ws.cell(offset+i+1, 1, each)
+        elif isinstance(v, (dict, OrderedDict)):
+            for i, key in enumerate(v.keys()):
+                cell = ws.cell(offset+1, i+1, key)
+                cell.font = pyxl.styles.Font(bold=True)
+            for i, row in enumerate(v.values()):
+                if isinstance(row, (list, tuple)):
+                    for j, c in enumerate(row):
+                        ws.cell(offset+j+2, i+1, c)
+                else:
+                    ws.cell(offset+i+1, 1, c)
+
+    workbook.remove(workbook['Sheet'])
+    workbook.save(path)