+++ hashcathelper/bloodhound.py
@@ -37,16 +37,18 @@ def get_driver(url):
     if not url:
         log.critical("No BloodHound URL given")
         exit(1)
-    regex = r'^bolt://(?P<user>[^:]+):(?P<password>.+)@'
+    regex = r'^bolt(?P<encrypted>s?)://(?P<user>[^:]+):(?P<password>.+)@'
     regex += r'(?P<host>.*):(?P<port>[0-9]+)$'
     m = re.match(regex, url)
 
-    user, password, host, port = m.groups()
+    encrypted, user, password, host, port = m.groups()
+    encrypted = (encrypted == 's')
 
     url = "bolt://%s:%s" % (host, port)
 
     log.debug("Connecting to %s..." % url)
-    driver = GraphDatabase.driver(url, auth=(user, password))
+    driver = GraphDatabase.driver(url, auth=(user, password),
+                                  encrypted=encrypted)
 
     return driver