+++ hashcathelper/subcommands/bloodhound.py
@@ -18,7 +18,7 @@ args.append(argument(
     dest='bloodhound_url',
     help="""
 URL to a Neo4j database containing BloodHound data. Format:
-bolt://<user>:<password>@<host>:<port>"""
+bolt[s]://<user>:<password>@<host>:<port>"""
 ))