+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix encoding when reading a file that probably comes from hashcat
+
 ## [0.1.3] - 2021-11-12
 
 ### Added