+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support `xlsx` format
 - Add parameter to change the minimum password length in the context of
   identifying short passwords
+- Include optional lookup in HIBP database in the detailed report
 
 ## [0.1.3] - 2021-11-12